Refactor repositories and services
- Refactor a MockDataRepository, rather than service, because its only
  used to load data into a ProjectRepository, and isn't a service itself.
- Move ProjectsDependency into a new Dependencies subpackage.
- Rename Repository into a ProjectsRepository within the repositories
  subpackage.
