+++ src/spherexportal/dependencies/__init__.py
@@ -1,0 +1,1 @@
+"""FastAPI dependencies for the SPHEREx Portal app."""
