+++ src/spherexportal/pages/handlers.py
@@ -9,7 +9,8 @@ from starlette.requests import Request
 from starlette.templating import Jinja2Templates, _TemplateResponse
 from structlog.stdlib import BoundLogger
 
-from spherexportal.repository import ProjectRepository, repository_dependency
+from spherexportal.dependencies.projects import projects_dependency
+from spherexportal.repositories.projects import ProjectRepository
 
 __all__ = ["router"]
 
@@ -24,14 +25,14 @@ templates = Jinja2Templates(directory=Path(__file__).parent / "templates")
 async def get_homepage(
     request: Request,
     logger: BoundLogger = Depends(logger_dependency),
-    repo: ProjectRepository = Depends(repository_dependency),
+    projects_repo: ProjectRepository = Depends(projects_dependency),
 ) -> _TemplateResponse:
-    context = {"request": request, "ssdc_ms": repo.ssdc_ms}
+    context = {"request": request, "ssdc_ms": projects_repo.ssdc_ms}
     return templates.TemplateResponse("index.html", context)
 
 
 @router.get("/__healthz")
 async def get_healthz(
-    repo: ProjectRepository = Depends(repository_dependency),
+    projects_repo: ProjectRepository = Depends(projects_dependency),
 ) -> PlainTextResponse:
     return PlainTextResponse("OK")
